Move time-related test from asdf/tests to asdf/tags/time
